Make ImuInitialize generic, just named Initialize
