+++ common/src/abstraction/mod.rs
@@ -1,5 +1,7 @@
+pub mod initialize;
+
+pub mod accelgyro;
 pub mod barometer;
-pub mod imu;
 pub mod magnetometer;
 pub mod motor;
 pub mod trigger;
