+++ common/src/drivers/imu/icm20948.rs
@@ -9,7 +9,8 @@ use icm20948_async::{
 pub use icm20948_async::{AccDlp, AccRange, AccUnit, Config, GyrDlp, GyrRange, GyrUnit};
 
 use crate::{
-    abstraction::imu::{AccelGyro, ImuInitialize},
+    abstraction::accelgyro::AccelGyro,
+    abstraction::initialize::Initialize,
     errors::SensorError,
     types::measurements::Imu6DofData,
     wrapped::{i2c::WrappedI2c, spi::WrappedSpi},
@@ -81,7 +82,7 @@ fn map_setup_err_spi<E: spi::Error>(error: SetupError<E>) -> SensorError {
 
 pub struct Icm209486DofI2c;
 
-impl<BUS> ImuInitialize for (Icm209486DofI2c, BUS)
+impl<BUS> Initialize for (Icm209486DofI2c, BUS)
 where
     BUS: i2c::I2c,
 {
@@ -116,7 +117,7 @@ where
 
 pub struct Icm209486DofSpi;
 
-impl<BUS> ImuInitialize for (Icm209486DofSpi, BUS)
+impl<BUS> Initialize for (Icm209486DofSpi, BUS)
 where
     BUS: spi::SpiDevice,
 {
